+++ packages/react-form/src/useFieldGroup.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { useState } from 'react'
-import { useStore } from '@tanstack/react-store'
+import { useSelector } from '@tanstack/react-store'
 import { FieldGroupApi, functionalUpdate } from '@tanstack/form-core'
 import { useIsomorphicLayoutEffect } from './useIsomorphicLayoutEffect'
 import type {
@@ -29,7 +29,7 @@ function LocalSubscribe({
   lens: AnyFieldGroupApi
   selector?: (state: FieldGroupState<any>) => FieldGroupState<any>
 }>): ReturnType<FunctionComponent> {
-  const data = useStore(lens.store, selector)
+  const data = useSelector(lens.store, selector)
 
   return <>{functionalUpdate(children, data)}</>
 }