+++ packages/vue-form/src/index.ts
@@ -1,5 +1,5 @@
 export * from '@tanstack/form-core'
-export { useStore } from '@tanstack/vue-store'
+export { useSelector, useStore } from '@tanstack/vue-store'
 export * from './useField'
 export * from './useForm'
 export * from './useFormGroup'