+++ packages/react-form/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @tanstack/react-form
 
+## 1.33.5
+
+### Patch Changes
+
+- Updated dependencies [[`146ba25`](https://github.com/TanStack/form/commit/146ba25205afe46c2a8bfc1f2ef82609255772dc)]:
+  - @tanstack/form-core@1.33.5
+
+## 1.33.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/form-core@1.33.4
+
+## 1.33.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/form-core@1.33.3
+
+## 1.33.2
+
+### Patch Changes
+
+- [#2240](https://github.com/TanStack/form/pull/2240) [`d756476`](https://github.com/TanStack/form/commit/d756476b3ce5e6618a062b929a4a256fa0f4d683) - Use fresh FieldApi in current render on name change
+
+- Updated dependencies []:
+  - @tanstack/form-core@1.33.2
+
+## 1.33.1
+
+### Patch Changes
+
+- [#2206](https://github.com/TanStack/form/pull/2206) [`df52a12`](https://github.com/TanStack/form/commit/df52a12c8772cb025bfba6773c5ff0a488211662) - Re-export `useSelector` from TanStack Store adapters and document migration from deprecated `useStore` (fixes [#2203](https://github.com/TanStack/form/issues/2203)).
+
+- Updated dependencies [[`3c57f6d`](https://github.com/TanStack/form/commit/3c57f6d4311f823ab39374417edb5fae544ad15a), [`7b3012b`](https://github.com/TanStack/form/commit/7b3012b03e706cc409c2952964145a11a300d7fe), [`741da6b`](https://github.com/TanStack/form/commit/741da6bb659a319acaa55292564999fcbbc90012), [`3c57f6d`](https://github.com/TanStack/form/commit/3c57f6d4311f823ab39374417edb5fae544ad15a), [`6cd477a`](https://github.com/TanStack/form/commit/6cd477a97df9612019d8a2d9c97f2e86b1b59256), [`887a129`](https://github.com/TanStack/form/commit/887a1293375a168452e82b0fb482841ec02ad77f)]:
+  - @tanstack/form-core@1.33.1
+
+## 1.33.1
+
+### Patch Changes
+
+- Re-export `useSelector` from `@tanstack/react-store` (fixes [#2203](https://github.com/TanStack/form/issues/2203)). `useStore` remains available but is deprecated; prefer `import { useSelector } from '@tanstack/react-form'`.
+
 ## 1.33.0
 
 ### Minor Changes